All sensors give data in engineering units. PWM Thread created which gives output on Servo 4.

+++ Pluto-chibios-2.5/src/Sensors/Magnetometer/hld/hmc5883.h
@@ -58,7 +58,7 @@
 #define OP_MODE_SINGLE		1
 #define OP_MODE_IDLE		2
 
-extern int32_t hmc_range;
+extern float hmc_range;
 void initialize_HMC(uint8_t average, uint8_t ODR, uint8_t Mode, uint8_t Gain, uint8_t OP_Mode);
 void read_hmc_values(void);
 void read_hmc_data(void);
